telegram_bot/requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
peewee==3.15.4
python-telegram-bot==13.15
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2022.7
tzlocal==4.2
psycopg2-binary==2.9.5
